+++ verik-compiler/src/main/kotlin/io/verik/compiler/core/common/Cardinal.kt
@@ -19,6 +19,8 @@ package io.verik.compiler.core.common
 object Cardinal {
 
     val UNRESOLVED = CardinalUnresolvedDeclaration
+    val FALSE = of(0)
+    val TRUE = of(1)
 
     fun of(value: Int): CardinalConstantDeclaration {
         return CardinalConstantDeclaration(value)