+++ CHANGELOG.md
@@ -8,7 +8,6 @@
 - Support tasks with return values.
 - `fork` and `join` functions.
 - Support x and z bit constants.
-- `IF` cardinal function.
 - Type parameter binding annotation `@Bind`.
 
 ### Changed
@@ -16,6 +15,7 @@
 - Turn off `enableLineDirective` by default.
 - Renamed `Logical` to `Optional`.
 - Rewrote constant expression handling.
+- Merged `Optional` with `Cardinal`.
 
 ## [0.1.11]
 ### Added